Fix casting problem in `hiff()`
The implementation of `hiff()` had `bits.len() as i64`, which could do Bad Things if `bits.len()` wouldn't fit in an `i64`. Clippy has started fussing about this, so I addressed the problem.

I added an assertion to `hiff()` that checks that we're going to be OK, which then means I don't need to worry about checking in `do_hiff()` and I can safely use `as i64` there. Hopefully that won't have too much performance impact.

I also made `do_hiff()` private instead of public since it's an implementation detail that shouldn't be visible to the outside world.