Create `BoolGenerator`
We had an `impl Generator<bool> for f64` which worked, but when I came back to it I found it quite confusing for a bit. It's also not obvious that booleans are the _only_ thing we'd want to generate given a floating point probability.

So I created a `struct BoolGenerator` and used it to implement `Generator<bool>`. It's perhaps wordier, but I think it's less confusing, especially for people new to the codebase.