Address casting warnings in `hiff()`
Updates to Clippy were complaining about `len as i64`, which was casting a `usize` to a `i64`. This had bugged me at the time, but there weren't any warnings so I let it go, which illustrates the importance of warnings like this.

So now this fixes that. I assert in `hiff()` that the length will fit in an `i64`, and then `do_hiff()` can safely convert that length and all sub-lengths to `i64` using `as i64`.