Modify `run_to_completion` to return a `Result` type
This modifies `(Push)State::run_to_completion` to return a `Result` type. It also updates `main` to use that new version of `run_to_completion`. In doing this we needed access to `ErrorSeverity` in `main()` so we added a `severity()` method and changed the visibility of `ErrorSeverity`.

I also (at the suggestion of Justus_Fluegel@Twitch) added pretty printing to the output in `main`.