Change default for `Mod` to 0
I'd just copied 1 down as the default from `ProtectedDivision`, but some thinking convinced me that 0 is the better answer since x % x == 0 for all x != 0.

It turns out that I'd also forgotten to change `ProtectedDivide` to `Mod` in the test, which cause the test to fail. This fixes that as well.