Add a `Debug` implementation for `PushInstruction`
This adds a custom `Debug` implementatin for `PushInstruction` that makes the printing of the instructions a lot more readable.

This changes variables so that they're just written as the variable name, and then `Bool` and `Int` instructions just have `Bool-` and `Int-` prefixes instead of the whole enum wrapper.

This follows up on a suggestion from Justus_Fluegel@Twitch.