Make `Error` an enum of `Recoverable` and `Fatal` variants
This changes `Error` to be an enum, with new types `RecoverableError` and `FatalError`. This allows us to return just one or the other error type when only one is possible, allowing the type system to help us out more. We did end up with some duplication in the implementation of some of the `Error` methods, but it wasn't bad.

We added several methods to `Error`, such as `is_recoverable()` and "getter"s like `state()` that simplified code that worked with `Error`.