Change `run_to_completion()` to return only a `FatalError`
Since `run_to_completion()` should "skip over" recoverable errors, the only kind of error that this should return is a `FatalError`. This is a nice advantage of the new enumerated type for `Error`; we can use the type system to confirm that we're returning the right error and don't need `unreachable!()` to say that things are impossible.