Add a `TryRecover` trait
The `TryRecover` trait provides the `try_recover()` method, which we implement on some specific `Result` types. The key idea is that if the `Result` is a `RecoverableError` error type, then we extract the state and return that as an `Ok()` variant, i.e., we "recover" from that error. If the `Result` is a `FatalError` type, then we return a `FatalError`.

The "magic" of this is arguably in line 148 where we convert an recoverable error into an `Ok`. The use of `Infallible` on lines 155-157 is also cool as it means that code like `PushState::run_to_completion` doesn't have to work about `RecoverableError`s since those will never be passed along. This allows us to avoid `match` logic in places like `PushState::run_to_completion()` in a nice way.

This was a suggestion from Justus_Fluegel@Twitch.