Embed a state transition diagram in the docs
Justus_Fluegel@Twitch generated a diagram of the `Builder` states using Excalidraw (with an addition by MizardX@Twitch), and then we figured out how to embed that into the relevant Rust docs comments.

I think we'll probably want to replace this with a figure generated with the `aquamarine` crate (suggested by esitsu@Twitch), but this gets us started.