Reformat `generate_builder.rs`; fix bug in `with_program`
Because of some trailing whitespace, Rust Analyzer wasn't properly reformatting this code, even though `cargo fmt --check` was all grumpy about it.

esitsu@Twitch suggested changing the VS Code settings to enable the "Trim Trailing Whitespace". That automagically found and removed the problematic trailing whitespace, after which Rust Analyzer was able to reformat the rest of the file.

The code generated for `with_program` reversed the given program before adding it to the `exec` stack, which turned out to be wrong. (Luckily the `run_simple_program` test highlighted this issue.) So we replaced the reversal with just extending the `exec` stack with the given program. As part of debugging this I added some comments to `run_simple_program` to clarify the intended stack state after each instruction is performed.