Fix the doctest for `Stack::try_extend`
This doctest didn't compile because of import problems. @JustusFluegel and esitsu@Twitch helped me work those problems out and we got this test to run and pass.

@JustusFluegel then showed me some nice tricks in writing doctests that I wasn't aware of. I didn't know, for example, that you could return `Ok` at the end of the doctest to tell the compiler that this function returns a `Result` type, which then allows you to used `?` instead of `unwrap()` everywhere. I think that helped remove a lot of clutter from the example code in a nice way.

They also showed me how I could use `#` to "hide" certain example lines (like the imports and that `Ok` line) so that the code still compiles and runs, but tidies up the code in the docs example so it's more focused on the relevant parts.

Very cool!

In the process of dealing with these, we changed `Stack` from being `pub(crate)` to just `pub`. Nobody could really remember why we'd made it `pub(crate)` (i.e., hidden it outside the `push_vm` crate), and making it `pub` allowed us to import it into the doctest code.