Get doctests in the state builder to pass
Two of the doctests in the state builder were marked as `ignore` and this unignores them and gets them to pass with the new code structure.

We also added a `Stack::max_stack_size()` method so we have access to the maximum size of the stack, which we use in one of these tests.