Tidy up doctests in `push-macros`
We chose to change one code block from `ignore` to `text` to reduce the number of ignored tests in the `cargo test` output.

I tried getting the other big doctest block to compile, but in the end it didn't work out because there are too many type dependencies on the `push` package that we don't access to (e.g., `Stack`). So we just left that as `ignore`.

We did change `exec` to be a `Stack` instead of a `Vec`, and added `?` to the appropriate calls in `main()`, having `main()` return a `Result` type.