Genericize `Error` and `Score`
This makes the `Error` and `Score` types generic in some type `T` to give us more flexibility.

This also make `TestResult` generic in two types, one for errors and one for scores. Not entirely sure if we'll ever use that, but it seemed a potentially nice piece of flexibility.

When switching from `i64`, which is `Clone + Copy` to generic `T`, we didn't want to assume the ability to `Clone` or `Copy`, so we changed summation over references to require `T: ToOwned` so we could use `.to_owned()` to convert references to owned types.

We also removed the `From<&T>` impls. They weren't being used, and getting them to work would require adding constraints on `T` that didn't seem necessary or desirable at this point. Callers should call `.copy()`/`.copied()`, or `.clone()`/`.cloned()`, as appropriate beforehand. See the `create_test_results...` tests for examples of this.

There were initially no tests for `PartialOrd` for `Score` or `Error`, so we added some.