Add GitHub ci (#22)
Add GitHub CI config

This is basically copied from https://doc.rust-lang.org/cargo/guide/continuous-integration.html with a few changes, mostly adding `clippy` and `fmt` checks.

I had to add the clippy and cargofmt components for this to work with `nightly` and `beta`. I'm not sure we _need_ (or even _want_) `nightly` or `beta`, but if there are issues there it could give us a heads up about things we'll need to transition to later. So I left them in for now.