Create an `push/examples` folder
I started this process by creating an `examples` folder in the `push` package and copying the `main.rs` file from the `push-cli` package over there. The goal is to create a set of examples showing how to use the `push` library to create and run PushGP experiments.

This required adding `clap` as a dev dependency for the `push` package. Adding it as a *dev* dependency means it won't get compiled in when people *use* the package unless they also choose to depend on it.