Add `Stack<f64>` to `PushState`
This created problems since `PushState` used to implement `Eq`, but `f64` doesn't implement `Eq`. To move the train forward we just removed the requirement that `PushState` derive `Eq` and `PartialEq`.

This also required dropping the requirement that `Stack` and `PushInstruction impl Eq`.

I think we'll want to at least revisit this for the instructions since we will want to be able to compute diversity metrics for Plushy genomes and Push programs, which will require some notion of equality. This gets us moving here, though, for now.