Convert `complex-regression` to use floats
This changes the implementation of `complex-regression` to use floats, using the `ordered_float` crate to wrap the `f64` values so that we can use things like selection.

I also stubbed out the add, subtract, multiply, and divide instructions in `float.rs`, but still need to implement `perform` for those instructions.