Exclude the `rust-lexicase` package
The `rust-lexicase` package is complicated to build because of the weird Python dependencies. Given that I'm not sure we're ever going to pursue that, and that if we did it should probably be in its own repo, I'm for not just excluding it from the list of workspaces to build and test.