Implement basic floating point Push instructions
This uses `OrderedFloat` to wrap all the floating point values so that things like `FloatInstruction` can implement `Eq` and `PartialEq`.

The implementation of the four basic arithmetic instructions are essentially identical to those for integers. The only real difference is that here we don't need to check for integer arithmetic errors (like overflow).

I suspect there's a generic version of instruction for numbers that can handle both integers and floats, although I haven't thought about what we've have to do with those errors.