Change Plushy genome to hold `PushGene`s
This introduces a new `PushGene` type that is either a `Close` or a `PushInstruction`, with a `From` implementation that converts `PushInstructions` to genes.

We also added a `GeneGenerator` structure that implements `Generator<PushGene>` with specified (and default) rates of `Close` genes.

Then the `Plushy` genome is changed to hold `PushGene`s instead of `PushInstructions`.

We also tidied up the tests to use the `.into()` methods to simplify the instruction lists.