Improve generalization of `Generator<Plushy>`
We ran into an error when generalizing the construction of generators that totally flummoxed me, but happily esitsu@Twitch figured it out. The problem was that our implementation of `Generator<Plushy> for CollectionGenerator<GeneGenerator>` was too specific. We only implemented it _specifically_ for `GeneGenerator`, when we really wanted to implement it for any type `GG` that implements the trait `Generator<PushGene>`. This was almost certainly me being overly specific at some point in the dim distant past (either 2 months ago or 8 months ago).

This fixes that.

Also did a search for implementations of `Generator<>` for `CollectionGenerator<>` and found a similar issue in `Generator<Bitstring>`, so I made a similar change there.