Add `FnScorer` trait and blanket implementations of `Scorer`
This introduces a new `FnScorer` struct that is just a wrapper around a type `T` that will (in practice) implmenet `Fn(&G) -> R` where `G` and `R` are, respectively, the thing being scored (usually a genome) and the type of the result (a score).

We then have an implementation of `Scorer` for `FnScorer` that just calls the wrapped function to do the scoring.

We also created a blanket implementation of `Scorer for &T` where `T` is an implementation of `Scorer`. This let's use pass references to `Scorers` when we need a `Scorer`.

The overall result is that we can use an `FnScorer` whenever we need a `Scorer`, and we can also use a `&Scorer` (or an `&FnScorer`) whenever we need a `Scorer`.