Update the `WithScorer` trait to use new `Scorer` changes
This updates the types of `with_scorer`, and adds a `with_scorer_fn`, that use the changes we made to the `Scorer` trait and the new `FnScorer` struct.

We ended up moving the generic `S` to the function instead of the struct, in part because `with_scorer_fn` doesn't use `S`, but instead needs `F`.