Update all the `main` functions
Many (all?) needed to be updated to take advantage of the new `FnScorer` and blanket `Scorer` implementations.

The `complex_regression::main()` also needed a bunch of changes to take advantage of the new `Case(s)` types.

We also created an `Of64` type so we didn't have to write `OrderedFloat<f64>` over and over again.