Make `Case` and `Cases` more flexible
This implements `From<(Input, Output)>` for `Case` and adds a `Case::new`. We then modified `Cases::add_case` to take `impl Into<Case<...>>` instead of explicit input and output values.

GitHub CoPilot also suggested a `From` the converts `Case` to `(Input, Output)`, so we added that as well.

We also added a `with_case` to support chaining of adding cases.