Draft problem Smallest
This is a start on the Smallest problem from PSB1. I don't think I have the training cases set up as they are in PSB1, and I don't have the full instruction set that was used for Smallest, but it does give us something that uses the full integer instruction set.