Implement `Exec::When`
This implements `Exec::None` and `Exec::When`, but doesn't have tests yet, which isn't great.

I got totally bogged down here in the design surrounding the error handling and `InstructionResult` holding a copy of the state in any `Err` variant. I had really expected this to be "easy" and it turned out to be a mess where I couldn't remember how our design worked. Ugh. I think we ended up in a reasonable place, but I think that things would have been a lot simpler if we could have just sprinkled in some `?`s in a few places.