Add `StackPush` trait
We realized today that there wasn't a `StackPush` trait similar to things like `StackDiscard`, and that seemed a non-trivial omission. So we renamed existing `StackPush` trait (which really did something different) to `PushOnto`, and created a `StackPush` trait that could be used in the same way as things like `StackDiscard`.

The changed a bunch of calls in some of the other instructions files.