Start drafting some `IfElse` tests
Didn't make a lot of progress here, but it's a start.

I have a test for when all the stacks are empty, but it fails because performing `IfElse` with empty stacks returns an `Err`.

I also started a `proptest` test, but there are a pile of warnings and import issues that are being difficult. I think I should probably move that to its own file.