Document and fix `When` instruction
This adds fairly detailed documentation for the `When` instruction that we think makes a reasonable starting point as a general instruction documentation template.

Among other things, this lists all five cases for the `When` instruction.

In doing this we decided that we should return an underflow error when both the boolean and exec stacks are empty. We then updated the relevant test to confirm that behavior, and then updated the `.perform()` method to get that updated test to pass.