Document and fix `Unless` instruction
This adds documentation for the `Unless` instruction following the template we generated for the `When` instruction.

We also decided that `Unless` should return a recoverable underflow error if both stacks are empty. The tests were modified to reflect that, and then the `.perform()` was updated to make the new tests pass.