Add `IfElse` documentation; rationalize `When` & `Unless`
This adds Rust doc comments for `IfElse`, and updates `When` and `Unless` so that their documentation matches this form.

The change was replacing the "Returns error" column with a "Success" column, and updating the emojis accordingly. We also added links to the error emojis that take you to the docs for the specific error being returned.