Rename `Score` and `Error` to `ScoreValue` and `ErrorValue`
The use of `Error` is confusing in the Rust world because of its many uses there. After _lots_ of discussion I chose to go with `ScoreValue` and `ErrorValue`, although not everyone was thrilled with those choices.

An interesting option is `performance::Error`, but that can't be enforced, which is awkward.

Other options included `Deviation` and `Transgression`. Neither of these are commonly used in the EC world. `Deviation` is potentially signed, where `Error` typically isn't, and `Transgression` seems more qualitative than quantitative.