Add associated `Error` types for `Operator`s
This adds an associated `Error` type for the `Operator` trait and changes the return type for `apply` to be a `std::result::Result` instead of an `anyhow::Result`.

This forced changes in anything that implemented `Operator`, as well as several things that acted on `Operator`s.

In several cases we had to add bounds saying that `anyhow::Error: From<X::Error>` to ensure that we could convert associated errors to `anyhow::Error`s. This (and other uses of `anyhow` in `ec_core`) should ultimately be removed as we transition to `miette` errors.