Duplicate the `example_runner` config for Median
I had incorrectly understood how the TOML syntax for `example_runner` config worked, and it looks like we need to duplicate everything for both Smallest and Median. Once I get around to changing `--num-generations` to `--max-generations` everywhere then we can remove all of this from `push/Cargo.toml`, but for now it's a bit messy.