Add specific `strum(to_string = …)` for Push `Push` instructions
We need to display the values being pushed when we use a `Push(T)` instruction for `BoolInstruction`, `IntInstruction`, or `FloatInstruction`. This uses the `#[strum(to_string = "Push({0})")]` annotation to get each of these to display with a form like `Float-Push(1)`.

This fails with the versions of `strum` and `strum-macros`, that we were using so I ran `cargo update` to update all the dependencies. This lead to a Clippy warning "multiple versions for dependency `heck`: 0.4.1, 0.5.0", which I think will cause the checks to fail on this PR.