Add tests for `to_string()` for `Push(T)` instructions
This adds proptests for `BoolInstruction`, `IntInstruction`, and `FloatInstruction`, checking that `to_string()` on each of the `Push(T)` instructions includes the value of the value being pushed. This isn't true with the default `#[derive(strum_macros::Display)]`, and these tests are meant to catch that if we regress later on.