Convert `IntInstruction::Abs` to new form with `Abs` type
This follows the pattern we used for `Negate` and creates a new `Abs` type which implements `Instruction`. The `IntInstruction::perform()` then just delegates `.perform()` to this new type when appropriate.

We moved the one test for `Abs` to this new module, and added a new test that specifically checks for the appropriate behavior on `i64::MIN`, which is the source of the bug.