Convert tournament size from `usize` to `NonZeroUsize`
This ensures that all tournaments are positive. It has the unfortunate side-effect of now requiring an `.unwrap()` or similar whenever constructing a `Tournament`. I added a `binary()` constructor that creates a binary tournament selector and handles the `unwrap()`, simplifying that fairly common use case. We could add additional tournament constructors for other common tournament sizes.