Address new Clippy warning about `ok_or`
There's a new Clippy warning in nightly about using `ok_or` when the argument is a function call. Using `ok_or` means that the function call will be evaluated regardless of the value of the `Option`. The recommendation is to use `ok_or_else` so that function call is handled lazily.