Use `filter` to clean up generation of distinct scores
In `tournament_size_2_pop_size_3` I was using an arguably opaque `mod 3` property to ensure that my three scores were distinct. This was in a perhaps misguided effort to avoid using `filter`, and after discussion with @JustusFluegel I was convinced to switch over to `filter`. That also removed the need for the `arithmetic_side_effects` expectation, which is nice.

Co-authored-by: Justus Fluegel <justusfluegel@gmail.com>