Add comments and doctests for `Cases`
This adds comments and doctests for all the methods in `Cases` and its various trait implementations.

I also re-ordered the various functions, trying to group them by:

   * Things that create sets of `Cases`
   * Properties of `Cases` (like `len()`)
   * Iterators over `Cases`

I also found the tests to be a bit cleaner with the addition of `.inputs()` and `outputs()`, which return iterators over the inputs and outputs for a set of `Cases`.

I removed two TODO comments as we've created issues for those.