Clean up `Cases` tests
This introduces the use to the `test_case` crate to simplify some testing in `with_target_fn.rs`.

We also cleaned up a bunch of tests, removing unnecessary calls to `.collect()` and better using iterators and `.eq()` on iterators.

We replaced the `iter.next(); iter.next()` pattern with `.eq()` as well.

This cleans up several Clippy warnings on doctests.