Add `DynSelect` to `Weighted`
This adds a new `DynSelect` trait that returns results with boxed `Error`s, allowing us to have arbitrary selectors and errors in `Weighted`. The boxing does, however, erase information about the specific kind of selector and error.

The implementation of `Selector` for `Box<S>` is no longer needed (`DynSelect` basically does this), so it was removed.