Start drafting recursive representation of weighted selectors
This is the beginning of implementing an idea proposed by @JustusFluegel. Instead of storing the selectors and weights in a `Vec`, this stores them in a recursive structure (essentially a hand-rolled linked list).

A key goal here is to avoid `Box`ing both the selectors and their associated errors so we don't "erase" those types. This will instead built quite complex nested types (so not tons of fun to look at if they're printed out), but they'll be complete, providing the compiler (and devs if they can make sense of them) detailed type info.

I wonder if we can use the new-ish feature that was released sometime in that past year (and whose name I don't remember) to improve the error messages that come from incorrect implementations of this trait. But first we need to finish this and see if it even works.