Add/improve tests for `selector` module
I added unit tests to this module, based largely on the doctests that we wrote earlier in the Twitch stream. This led to several improvements in the doctests, and I ended up largely rewriting them.

They will need some proofreading, but I think this is a substantial improvement, and a nice model to use in the other operators.