Address numerous code review suggestions
@JustusFluegel provided a lot of excellent suggestions, that then led to these changes. Their suggestions largely fell into three categories:

- Grouping imports and a few other formatting bits
- Using arrays (e.g., `[u8; 4]`) instead of vectors whenever reasonable
- Using `?` instead of `.unwrap()` in the doctest code

In going through the vector->array chances, I realized that some of the examples were more generic than others, and I moved to have everything take on the more generic approach. In the process, I introduced a `Genome<T>` type, which will make it a lot easier to change that type if we decide to go back to vectors, or (more likely) we move to `DerefMut<[T]>`.

Switching from `.unwrap()` to `?` was a little more complex than I expected. I needed to return an `Ok::<(), anyhow::Error>(())` at the end of each doctest block so that the implicit wrapping function would have a return type that allows the use of `?`.