Add short doc comments to `IntInstruction` enum
The generated docs for `IntInstruction` just had a big list of all the enum variants, so I added a short doc comment to the first four.

In doing this, I discovered that the more comprehensive documentation in files like `abs.rs` wouldn't be visible without the `document-private-items` flag because we were "hiding" those. We clearly want that documentation to be generally visible, so I changed the visibility so that these types wouldn't be private. I'm not sure if that was the best plan or not, though.