Add `AsciiFromWrappingFloat`
This adds the ability to create an ASCII character from the top of the `OrderedFloat<f64>` stack.

The current implementation cases the `f64` to an `i128`, and then uses `euclid_rem(128)` to map that to `0..128`, before converting to a character.

I thought about using `euclid_rem()` in `f64` space, but there are apparently (rare) circumstances where that could return 128, so I chose to convert to `i128` first. Not 100% sure that was the best move, though.