Add `Dup` command
This still needs tests.

It also essentially mimics what's in `DupBlock`, but is more general, so we should delete `DupBlock`.

One thing I'm not clear on is whether we should have a `Common` enum with a `Dup(Dup)` variant like we (e.g.) have `Abs(Abs)` variants for `IntInstruction`? What's the actual value of having the instruction enums with the (rather odd) `Abs(Abs)` variants? Does it provide extensibility? Does it make it easier to group instructions? (I'm pretty sure it does do that, so maybe that's reason enough.)