Add unit and property-based tests for `Dup` instruction
This is really similar to `Exec::DupBlock`, but not quite the same because `DupBlock` has an implied "open" after it. Maybe we could have just a single instruction, though, since (I think) that the implied open is ignored on any non-exec stack.

We also need to move `Dup` into its own file (`dup.rs`) and create a `Common` enum for instructions like this.