Fix a bug in `StackDepth` tests
In the test `correct_stack_sizes_for_bool` I was setting the max stack size of the number of values being pushed onto the boolean stack. This didn't take into account the possibility of the number of values on the bool stack being 0, which caused the max stack size to be 0, which led to an error when we tried to push the stack size onto the int stack.

Luckily proptest caught this for us.