Add common instructions for the exec stack
This forced us to implement the `NumOpens` trait for the common instructions.

We also had to wrap the `PushValue<PushProgram>` in a `Box` in `ExecInstruction` to avoid an recursive enum definition in:

PushProgram -> PushInstruction -> ExecInstruction -> PushValue -> PushProgram

Luckily Rust's auto-deref make this largely transparent, where the `perform` implementation automatically dereferences the `Box`, giving us the contents for free.