Use `Ord::clamp` in implementation of `Clamp`
This uses the standard library implementation of `.clamp()` instead of home grown one. The `Ord::clamp()` implementation checks that `min <= max`, which mine didn't, and that required updating the instruction and the tests. We went with just taking the smaller of the relevant stacks to be the `min` and the larger to be the `max` to provide the most flexibility for evolution.

Thanks to @JustusFluegel for pointing out that this exists.