Specify `max_instruction_steps` in benchmark tests
It turns out that because a numeric fields defaults to a value of zero, if you don't call `set_max_instruction_steps()` you get a limit of 0 steps, and no instructions in your program won't be evaluated. This will typically mean that you _have_ to provide a value for `max_instruction_steps` when creating a `PushState`, which is further evidence that we should add that to the macro that builds states.