Implement generic `NotEqual` instruction
This is basically a copy of `Equal` with the appropriate changes.

Similar to `Equal`, the tests the try to use two different stacks are commented out because it's just not clear how to even get them to compile.

I realized in going through this that the comments on `Equal` were a bit of a mess (I'd copied them from elsewhere and didn't finish modifying them), so I cleaned that up as well.