Fix tests for `Equal` and `NotEqual`
Adding the required `.with_instruction_step_limit()` to the state builder broke all the tests for these operators, but that was easy to fix.

I also removed the commented out tests for stacks of different types since we're not likely to see that anytime soon, and I renamed a few tests to try to more clearly indicate what they were testing.