Add "integration" tests based on Median problem
The bugs with the integer and float comparison operators were turned up by trying to make sense of an evolved "solution" to the Median problem that didn't make sense.

I created this file to help me debug the problem, and then to capture the issue in case I regress on this at some future point.

I put this in the `tests/` folder since it wasn't clear where else it would go that would make more sense.