Make anonymous lifetime explicitly in `Bitstring` iterators
An update to Clippy adds a warning that "lifetime flowing from input to output with different syntax can be confusing" and suggests adding explicitly anonymous lifetimes to the return types for `Bitstring::iter()` and `Bitstring::iter_mut()`. This does that.

This also resolves a similar issue for `Cases::iter_mut()`.