Add a (commented out) 2-point XO impl with `choose_multiple_fill()`
It turns out that using `choose_multiple_fill()` is _really_ slow compared to the previous implementation (60x slower). Unfortunately the previous implementation isn't actually *correct* because it doesn't ensure that the two crossover points are distinct.

For the moment the correct implementation is commented out, and we'll have to figure out how we're going to address this.