Fix `TwoPointXo` implementation
This implementation was broken because it allowed selection of 0 as a crossover point and it allowed for `first == second`, both of which leads to an empty segment.

We fixed this by changing the start of the range to 1 and essentially unrolling one loop of Floyd's selection algorithm to ensure distinct crossover points.